+++ sdk/eventhub/Azure.Messaging.EventHubs/CHANGELOG.md
@@ -6,8 +6,8 @@
 
 ### Breaking Changes
 
-- Change `ActivitySource` name used to report message activity from `Azure.Messaging.EventHubs.EventHubs` to `Azure.Messaging.EventHubs.EventHubMessage`
-  and message `Activity` name from `EventHubs.Message` to `EventHubMessage.Create`
+- Change `ActivitySource` name used to report message activity from `Azure.Messaging.EventHubs.EventHubs` to `Azure.Messaging.EventHubs.Message`
+  and message `Activity` name from `EventHubs.Message` to `Message`
 
 ### Bugs Fixed